Extract *Selector(s) and Iconbutton into their own files
`Selector` did receive a small dose of love. Current does NOT utilize
react hooks. In our case it’s a waste of resources, this time around.
